Fix others warnings.

+++ include/TheLostGirl/FixtureRoles.h
@@ -4,7 +4,7 @@
 /// \file FixtureRoles.h
 
 /// Enumeration of role of some fixture (foot, head, ...).
-enum class FixtureRole : unsigned int
+enum class FixtureRole : long unsigned int
 {
 	None  = 0,      ///< Role that correspond to none role.
 	Foot  = 1 << 0, ///< Role for fixtures that observes the collisions with the ground.
@@ -17,7 +17,7 @@ enum class FixtureRole : unsigned int
 /// \return True if the given fixture has all its roles flags that match with the given flags, false otherwise.
 inline bool fixtureHasRole(b2Fixture* fixture, FixtureRole role)
 {
-	return fixture->GetUserData() and (reinterpret_cast<unsigned int>(fixture->GetUserData()) & static_cast<unsigned int>(role));
+	return fixture->GetUserData() and (reinterpret_cast<long unsigned int>(fixture->GetUserData()) & static_cast<long unsigned int>(role));
 }
 
 #endif//FIXTUREROLES_H
